backend/requirements.txt
# Requirements for Python 3.8+ (Latest versions supported)
# Pure Python packages - no compilation required
Flask==3.1.0
Flask-CORS==4.0.1
opencv-python==4.9.0.80
numpy==1.26.2
Pillow==10.1.0
